fix launch problem under linux

+++ s0il/bundled/lua/luac.c
@@ -316,7 +316,7 @@ static void PrintConstant(const Proto* f, int i)
   case LUA_VNUMFLT:
 	{
 	char buff[100];
-	sprintf(buff,LUA_NUMBER_FMT,fltvalue(o));
+	sprintf(buff,LUA_NUMBER_FMT,(double)fltvalue(o));
 	printf("%s",buff);
 	if (buff[strspn(buff,"-0123456789")]=='\0') printf(".0");
 	break;
